fix: move dnd config to separate component
